% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggoeg.R
\name{theme_goeg}
\alias{theme_goeg}
\title{ggplot2 theme parameters that modify plots}
\usage{
theme_goeg(...)
}
\arguments{
\item{...}{additional ggplot theme parameters that overwrite or supplement the GÖG theme}
}
\value{
an (incomplete) ggplot2 theme
}
\description{
ggplot2 theme parameters that modify plots
}
